+++ apps/server/src/shared/domain/index.ts
@@ -1,4 +1,4 @@
 export * from './entity';
 export * from './interface';
 export * from './types';
-export * from './vo';
+// export * from './vo';
