--- apps/server/src/shared/domain/vo/index.ts
@@ -1,1 +1,0 @@
-export * from './task-with-status.vo';
