--- apps/server/src/shared/domain/vo/task-with-status.vo.ts
@@ -1,20 +1,0 @@
-import { Task } from '../entity/task.entity';
-
-interface ITaskStatus {
-	submitted: number;
-	maxSubmissions: number;
-	graded: number;
-	isDraft: boolean;
-	isSubstitutionTeacher: boolean;
-}
-
-export class TaskWithStatusVo {
-	task!: Task;
-
-	status!: ITaskStatus;
-
-	constructor(task: Task, status: ITaskStatus) {
-		this.task = task;
-		this.status = status;
-	}
-}
